Update message to indicate what actually failed
This is a bit confusing since we fail the workflow on this step instead of the actual step that had the issue. We still want this to run so we can create the failed deployment on Github, but the real error is in the previous step so hoping this makes that clearer.